fix:오타 라우터 이름 수정
